+++ libnavui-androidauto/src/main/java/com/mapbox/androidauto/action/MapboxActionProvider.kt
@@ -1,8 +1,0 @@
-package com.mapbox.androidauto.action
-
-import androidx.car.app.Screen
-import androidx.car.app.model.Action
-
-fun interface MapboxActionProvider {
-    fun getAction(screen: Screen): Action
-}