+++ libnavui-androidauto/src/main/java/com/mapbox/androidauto/freedrive/FreeDriveActionStrip.kt
@@ -6,6 +6,7 @@ import androidx.car.app.model.ActionStrip
 import androidx.car.app.model.CarIcon
 import androidx.core.graphics.drawable.IconCompat
 import com.mapbox.androidauto.R
+import com.mapbox.androidauto.feedback.ui.CarFeedbackAction
 import com.mapbox.androidauto.screenmanager.MapboxScreen
 import com.mapbox.androidauto.screenmanager.MapboxScreenManager
 
@@ -17,22 +18,14 @@ class FreeDriveActionStrip(
      */
     fun builder() = ActionStrip.Builder()
         .addAction(buildSettingsAction())
-        .addAction(buildFreeDriveFeedbackAction())
+        .addAction(buildFeedbackAction())
         .addAction(buildSearchAction())
         .addAction(buildFavoritesAction())
 
-    private fun buildFreeDriveFeedbackAction() = Action.Builder()
-        .setIcon(
-            CarIcon.Builder(
-                IconCompat.createWithResource(screen.carContext, R.drawable.mapbox_car_ic_feedback)
-            ).build()
-        )
-        .setOnClickListener {
-            MapboxScreenManager.push(MapboxScreen.FREE_DRIVE_FEEDBACK)
-        }
-        .build()
-
-    private fun buildSettingsAction() = Action.Builder()
+    /**
+     * Action to open settings screen.
+     */
+    fun buildSettingsAction() = Action.Builder()
         .setIcon(
             CarIcon.Builder(
                 IconCompat.createWithResource(
@@ -46,7 +39,16 @@ class FreeDriveActionStrip(
         }
         .build()
 
-    private fun buildSearchAction() = Action.Builder()
+    /**
+     * Action to provide feedback for the free drive state.
+     */
+    fun buildFeedbackAction(): Action = CarFeedbackAction(MapboxScreen.FREE_DRIVE_FEEDBACK)
+        .getAction(screen)
+
+    /**
+     * Action to open search screen.
+     */
+    fun buildSearchAction(): Action = Action.Builder()
         .setIcon(
             CarIcon.Builder(
                 IconCompat.createWithResource(
@@ -60,7 +62,10 @@ class FreeDriveActionStrip(
         }
         .build()
 
-    private fun buildFavoritesAction() = Action.Builder()
+    /**
+     * Action to open favorite places screen.
+     */
+    fun buildFavoritesAction(): Action = Action.Builder()
         .setTitle(screen.carContext.resources.getString(R.string.car_action_search_favorites))
         .setOnClickListener {
             MapboxScreenManager.push(MapboxScreen.FAVORITES)